add common folder for integration tests
