+++ debug_utils/sierra-emu/tests/corelib.rs
@@ -17,11 +17,13 @@ use cairo_lang_test_plugin::{
     test_config::{PanicExpectation, TestExpectation},
     test_plugin_suite, TestCompilation, TestsCompilationConfig,
 };
+use common::jitvalue_to_felt;
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
 use sierra_emu::{run_program, EntryPoint, ProgramTrace, Value};
-use starknet_crypto::Felt;
 use tracing::{error, info};
 use tracing_test::traced_test;
+
+mod common;
 
 #[test]
 #[traced_test]
@@ -103,13 +105,12 @@ fn trace_to_run_result(trace: ProgramTrace) -> RunResultValue {
         outer_value @ Value::Enum {
             index,
             payload,
-            debug_name: _,
+            debug_name,
             ..
         } => {
-            //let debug_name = debug_name.as_ref().expect("missing debug name");
-            if *index == 0
-            // debug_name.starts_with("core::panics::PanicResult::")
-            //     || debug_name.starts_with("Enum<ut@core::panics::PanicResult::")
+            let debug_name = debug_name.as_ref().expect("missing debug name");
+            if debug_name.starts_with("core::panics::PanicResult::")
+                || debug_name.starts_with("Enum<ut@core::panics::PanicResult::")
             {
                 let is_success = *index == 0;
 
@@ -161,7 +162,7 @@ fn assert_test_expectation(
                 error!("test {}: {}", name, err_msg);
                 success = false;
             }
-            info!("test: {}", name);
+            info!("test {} ... OK", name);
         }
         RunResultValue::Panic(e) => match expectation {
             TestExpectation::Success => {
@@ -176,100 +177,13 @@ fn assert_test_expectation(
                         error!("test {}: {}", name, err_msg);
                         success = false;
                     }
-                    info!("test {}", name);
+                    info!("test {} ... OK", name);
                 }
             }
         },
     }
 
     success
-}
-
-/// Convert a Value to a felt.
-pub fn jitvalue_to_felt(value: &Value) -> Vec<Felt> {
-    let mut felts = Vec::new();
-    match value {
-        Value::Array { data, .. } | Value::Struct(data) => {
-            data.iter().flat_map(jitvalue_to_felt).collect()
-        }
-        Value::BoundedInt { value, .. } => vec![value.into()],
-        Value::Bytes31(bytes) => vec![*bytes],
-        Value::BuiltinCosts(costs) => vec![
-            costs.r#const.into(),
-            costs.pedersen.into(),
-            costs.bitwise.into(),
-            costs.ecop.into(),
-            costs.poseidon.into(),
-            costs.add_mod.into(),
-            costs.mul_mod.into(),
-        ],
-        Value::CircuitModulus(value) => vec![value.into()],
-        Value::Circuit(data) | Value::CircuitOutputs(data) => data.iter().map(Felt::from).collect(),
-        Value::EcPoint { x, y } => {
-            vec![*x, *y]
-        }
-        Value::EcState { x0, y0, x1, y1 } => {
-            vec![*x0, *y0, *x1, *y1]
-        }
-        Value::Enum {
-            index,
-            payload,
-            debug_name,
-            ..
-        } => {
-            if let Some(debug_name) = debug_name {
-                if debug_name == "core::bool" {
-                    vec![(*index == 1).into()]
-                } else {
-                    let mut felts = vec![(*index).into()];
-                    felts.extend(jitvalue_to_felt(payload));
-                    felts
-                }
-            } else {
-                // Assume its a regular enum.
-                let mut felts = vec![(*index).into()];
-                felts.extend(jitvalue_to_felt(payload));
-                felts
-            }
-        }
-        Value::Felt(felt) => vec![*felt],
-        Value::FeltDict { data, .. } => {
-            for (key, value) in data {
-                felts.push(*key);
-                let felt = jitvalue_to_felt(value);
-                felts.extend(felt);
-            }
-
-            felts
-        }
-        Value::FeltDictEntry {
-            key: data_key,
-            data,
-            ..
-        } => {
-            felts.push(*data_key);
-
-            for (key, value) in data {
-                felts.push(*key);
-                let felt = jitvalue_to_felt(value);
-                felts.extend(felt);
-            }
-
-            felts
-        }
-        Value::I8(x) => vec![(*x).into()],
-        Value::I16(x) => vec![(*x).into()],
-        Value::I32(x) => vec![(*x).into()],
-        Value::I64(x) => vec![(*x).into()],
-        Value::I128(x) => vec![(*x).into()],
-        Value::U8(x) => vec![(*x).into()],
-        Value::U16(x) => vec![(*x).into()],
-        Value::U32(x) => vec![(*x).into()],
-        Value::U64(x) => vec![(*x).into()],
-        Value::U128(x) => vec![(*x).into()],
-        Value::U256(x, y) => vec![(*x).into(), (*y).into()],
-        Value::Unit | Value::Uninitialized { .. } => vec![0.into()],
-    }
 }
 
 fn compile_tests(
